Moved tracking functions to helpers
